Further test refinements
