+++ test/Microsoft.Azure.Devices.Edge.Test/helpers/ManualProvisioningFixture.cs
@@ -5,91 +5,6 @@ namespace Microsoft.Azure.Devices.Edge.Test.Helpers
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Devices.Edge.Test.Common;
-    using NUnit.Framework;
-
-    public class X509ManualProvisioningFixture : ManualProvisioningFixture
-    {
-        public X509Thumbprint Thumbprint { get; set; }
-        public EdgeDevice Device { get; set; }
-
-        public X509ManualProvisioningFixture()
-            : base()
-        {
-        }
-
-        [OneTimeSetUp]
-        public async Task X509ProvisionEdgeAsync()
-        {
-            await Profiler.Run(
-                async () =>
-                {
-                    using (var cts = new CancellationTokenSource(Context.Current.SetupTimeout))
-                    {
-                        CancellationToken token = cts.Token;
-                        DateTime startTime = DateTime.Now;
-
-                        this.Thumbprint = this.CreateSelfSignedCertificateThumbprint();
-
-                        EdgeDevice device = await EdgeDevice.GetOrCreateIdentityAsync(
-                        Context.Current.DeviceId + "-x509",
-                        this.iotHub,
-                        AuthenticationType.SelfSigned,
-                        this.Thumbprint,
-                        token);
-
-                        Context.Current.DeleteList.TryAdd(device.Id, device);
-
-                        this.Device = device;
-
-                        await this.ManuallyProvisionEdgeAsync(device, startTime, token);
-                    }
-                },
-                "Completed edge manual provisioning");
-        }
-
-        private X509Thumbprint CreateSelfSignedCertificateThumbprint()
-        {
-            return new X509Thumbprint()
-            {
-                PrimaryThumbprint = "9991572f0a02bdc7c89fc032b95d79aca18ef7a3",
-                SecondaryThumbprint = "9991572f0a02bdc7c89fc032b95d79aca18ef7a4"
-            };
-        }
-    }
-
-    public class SasManualProvisioningFixture : ManualProvisioningFixture
-    {
-        public SasManualProvisioningFixture()
-            : base()
-        {
-        }
-
-        [OneTimeSetUp]
-        public async Task SasProvisionEdgeAsync()
-        {
-            await Profiler.Run(
-                async () =>
-                {
-                    using (var cts = new CancellationTokenSource(Context.Current.SetupTimeout))
-                    {
-                        CancellationToken token = cts.Token;
-                        DateTime startTime = DateTime.Now;
-
-                        EdgeDevice device = await EdgeDevice.GetOrCreateIdentityAsync(
-                        Context.Current.DeviceId,
-                        this.iotHub,
-                        AuthenticationType.Sas,
-                        null,
-                        token);
-
-                        Context.Current.DeleteList.TryAdd(device.Id, device);
-
-                        await this.ManuallyProvisionEdgeAsync(device, startTime, token);
-                    }
-                },
-                "Completed edge manual provisioning");
-        }
-    }
 
     public class ManualProvisioningFixture : BaseFixture
     {
